requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.5.6
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.5b1
bleach==3.3.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
flake8==3.9.2
GPUtil==1.4.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lz4==3.1.3
MarkupSafe==2.0.1
matplotlib==3.2.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.1
numpy==1.21.0
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==0.13.1
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pyflakes==2.3.1
Pygments==2.9.0
pylint==2.8.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
pyyaml==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
soupsieve==2.2.1
terminado==0.10.0
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==4.4
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
